+++ .eslintrc
@@ -16,13 +16,6 @@
     "comma-dangle": [2, "never"],
     "max-len": 0,
     "no-shadow": 0,
-    "arrow-body-style": 0,
-    "global-require": 0,
-    "no-use-before-define": 0,
-    "no-unused-expressions": 0,
-    "no-confusing-arrow": 0,
-    "no-unused-vars": 0,
-    "no-constant-condition": 0,
     "object-curly-newline": 0,
     "import/prefer-default-export": 0,
     "import/no-extraneous-dependencies": 0,
